# [WIP] Speedup data loading for tensorflow_stub by threading and caching