--- a/tensorboard/backend/event_processing/io_wrapper.py
+++ b/tensorboard/backend/event_processing/io_wrapper.py
@@ -31,6 +31,8 @@ logger = tb_logging.get_logger()
 
 _ESCAPE_GLOB_CHARACTERS_REGEX = re.compile("([*?[])")
 
+directory_cache = {}
+use_bfs_and_build_cache_MT = True
 
 def IsCloudPath(path):
     return (
@@ -85,9 +87,12 @@ def ListDirectoryAbsolute(directory):
 
     The paths are absolute.
     """
-    return (
-        os.path.join(directory, path) for path in tf.io.gfile.listdir(directory)
-    )
+    if len(directory_cache) > 0:
+        return directory_cache[directory]
+    else:
+        return (
+            os.path.join(directory, path) for path in tf.io.gfile.listdir(directory)
+        )
 
 
 def _EscapeGlobCharacters(path):
@@ -162,6 +167,59 @@ def ListRecursivelyViaGlobbing(top):
         current_glob_string = os.path.join(current_glob_string, "*")
         level += 1
 
+import queue
+import time
+import os
+import threading
+class listdir_bfs:
+    def __init__(self, baseDir, list_dir_function=None):
+        # list_dir_function = listdirWithDelay
+        self.return_structure = []
+        self.list_dir_function = list_dir_function
+        start = time.time()
+        list_dir_function(".")
+        self.max_idle_time_in_sec = time.time() - start
+        self.max_idle_time_in_sec = self.max_idle_time_in_sec * 5
+        self.unwalked_dirs = queue.Queue()
+        self.unwalked_dirs.put(baseDir)
+        self.threads = []
+        for i in range(20):
+            t = threading.Thread(name="thread"+str(i), target=self.worker)
+            self.threads.append(t)
+            t.start()
+
+        for t in self.threads:
+            t.join()
+
+    def worker(self):
+        """thread worker function"""
+        if "last_job_time" not in locals():
+            last_job_time = time.time()
+        while True:
+            try:
+                current = self.unwalked_dirs.get(False)
+                last_job_time = time.time()
+            except queue.Empty:
+                if time.time() - last_job_time > self.max_idle_time_in_sec : # and last_job_time > 0:
+                    # print('Worker '+threading.current_thread().getName() + " is stopped")
+                    break
+                # print('Worker '+threading.current_thread().getName() + " slept for 0.1s")
+                time.sleep(0.1)
+                continue
+
+            children = self.list_dir_function(current)
+            file_paths = []
+            dir_paths = []
+            for c in children:
+                fullpath = os.path.join(current, c)
+                # check symbolic link?
+                if os.path.isdir(fullpath):
+                    self.unwalked_dirs.put(fullpath)
+                    dir_paths.append(fullpath)
+                else:
+                    file_paths.append(fullpath)
+            self.return_structure.append((current, tuple(dir_paths), tuple(file_paths)))
+            # print(self.return_structure)
 
 def ListRecursivelyViaWalking(top):
     """Walks a directory tree, yielding (dir_path, file_paths) tuples.
@@ -180,11 +238,22 @@ def ListRecursivelyViaWalking(top):
     Yields:
       A (dir_path, file_paths) tuple for each directory/subdirectory.
     """
-    for dir_path, _, filenames in tf.io.gfile.walk(top, topdown=True):
-        yield (
-            dir_path,
-            (os.path.join(dir_path, filename) for filename in filenames),
-        )
+    if use_bfs_and_build_cache_MT:
+        global directory_cache
+
+        xx = listdir_bfs(top, list_dir_function=tf.io.gfile.listdir)
+        directory_cache = {}
+        
+        for key, dirs, files in xx.return_structure:
+            directory_cache[key] = dirs+files
+        for dir_path, _,  filename in xx.return_structure:
+            yield (dir_path, filename)
+    else:
+        for dir_path, _, filenames in tf.io.gfile.walk(top, topdown=True):
+            yield (
+                dir_path,
+                (os.path.join(dir_path, filename) for filename in filenames),
+            )
 
 
 def GetLogdirSubdirectories(path):