* Motivation for features / changes
As described in #4023 , `tf.io.gfile.listdir` may suffer from network latency. This is a POC that caching the directory structure and travel the with multiple threads can speed it up. 
* Technical description of changes
  1a. To simulate the latency, I inserted the `time.sleep(0.1)` before each os.listdir() is called.
  1b. Added some print statement to compare the time spent.
  2. The original implementation of walk() is depth-first traversal, I implemented a `listdir_bfs` class so that the traversal can be concurrent.
  3. The traversal result is cached so that tensorboard reloading is very fast.
  4. [TODO] Find a proper location for `listdir_bfs` class and expose the `use_bfs_and_build_cache_MT` flag.
  5. [TODO] Unit test.
  6. [TODO] determine the number of concurrent threads, and the timeout for `listdir()`; expose the settings to the launch args.

* Detailed steps to verify changes work correctly (as executed by you)

`bazelisk run tensorboard -- --logdir runs`

```
with `use_bfs_and_build_cache_MT = True`
AddRunsFromDirectory done. Load took 2.502 secs
TensorBoard done reloading. Load took 80.473  <- first time load, slow because the deserialization of event file? This is not parallelized on my PC.
multiplexer.AddRunsFromDirectory(path, name) completed
AddRunsFromDirectory done. Load took 2.499 secs
TensorBoard done reloading. Load took 0.017secs
...
with `use_bfs_and_build_cache_MT = False`
AddRunsFromDirectory done. Load took 31.620 secs
TensorBoard done reloading. Load took 98.740 secs
AddRunsFromDirectory done. Load took 31.782 secs
TensorBoard done reloading. Load took 10.455 secs
...
```

  <details><summary>Testing folder structure</summary>
<p>

```bash
runs
├── a
│   ├── Aug05_09-09-17_twhuang-mbp
│   │   ├── 00000
│   │   │   └── default
│   │   │       ├── metadata.tsv
│   │   │       ├── sprite.png
│   │   │       └── tensors.tsv
│   │   ├── 00001
│   │   │   └── noMetadata
│   │   │       └── tensors.tsv
│   │   ├── 00002
│   │   │   └── default
│   │   │       ├── metadata.tsv
│   │   │       ├── sprite.png
│   │   │       └── tensors.tsv
│   │   ├── data
│   │   │   └── scalar_group
│   │   │       ├── arctanx
│   │   │       │   └── events.out.tfevents.1596589757.twhuang-mbp
│   │   │       ├── xcosx
│   │   │       │   └── events.out.tfevents.1596589757.twhuang-mbp
│   │   │       └── xsinx
│   │   │           └── events.out.tfevents.1596589757.twhuang-mbp
│   │   ├── events.out.tfevents.1596589757.twhuang-mbp
│   │   └── projector_config.pbtxt
│   ├── Aug05_09-09-36_twhuang-mbp
│   │   ├── 00000
│   │   │   └── default
│   │   │       ├── metadata.tsv
│   │   │       ├── sprite.png
│   │   │       └── tensors.tsv
│   │   ├── 00001
│   │   │   └── noMetadata
│   │   │       └── tensors.tsv
│   │   ├── 00002
│   │   │   └── default
│   │   │       ├── metadata.tsv
│   │   │       ├── sprite.png
│   │   │       └── tensors.tsv
│   │   ├── data
│   │   │   └── scalar_group
│   │   │       ├── arctanx
│   │   │       │   └── events.out.tfevents.1596589776.twhuang-mbp
│   │   │       ├── xcosx
│   │   │       │   └── events.out.tfevents.1596589776.twhuang-mbp
│   │   │       └── xsinx
│   │   │           └── events.out.tfevents.1596589776.twhuang-mbp
│   │   ├── events.out.tfevents.1596589776.twhuang-mbp
│   │   └── projector_config.pbtxt
│   ├── Aug05_09-09-56_twhuang-mbp
│   │   ├── 00000
│   │   │   └── default
│   │   │       ├── metadata.tsv
│   │   │       ├── sprite.png
│   │   │       └── tensors.tsv
│   │   ├── 00001
│   │   │   └── noMetadata
│   │   │       └── tensors.tsv
│   │   ├── 00002
│   │   │   └── default
│   │   │       ├── metadata.tsv
│   │   │       ├── sprite.png
│   │   │       └── tensors.tsv
│   │   ├── data
│   │   │   └── scalar_group
│   │   │       ├── arctanx
│   │   │       │   └── events.out.tfevents.1596589796.twhuang-mbp
│   │   │       ├── xcosx
│   │   │       │   └── events.out.tfevents.1596589796.twhuang-mbp
│   │   │       └── xsinx
│   │   │           └── events.out.tfevents.1596589796.twhuang-mbp
│   │   ├── events.out.tfevents.1596589796.twhuang-mbp
│   │   └── projector_config.pbtxt
│   ├── Aug05_09-10-15_twhuang-mbp
│   │   ├── 00000
│   │   │   └── default
│   │   │       ├── metadata.tsv
│   │   │       ├── sprite.png
│   │   │       └── tensors.tsv
│   │   ├── 00001
│   │   │   └── noMetadata
│   │   │       └── tensors.tsv
│   │   ├── 00002
│   │   │   └── default
│   │   │       ├── metadata.tsv
│   │   │       ├── sprite.png
│   │   │       └── tensors.tsv
│   │   ├── data
│   │   │   └── scalar_group
│   │   │       ├── arctanx
│   │   │       │   └── events.out.tfevents.1596589815.twhuang-mbp
│   │   │       ├── xcosx
│   │   │       │   └── events.out.tfevents.1596589815.twhuang-mbp
│   │   │       └── xsinx
│   │   │           └── events.out.tfevents.1596589815.twhuang-mbp
│   │   ├── events.out.tfevents.1596589815.twhuang-mbp
│   │   └── projector_config.pbtxt
│   └── Aug05_09-10-34_twhuang-mbp
│       ├── 00000
│       │   └── default
│       │       ├── metadata.tsv
│       │       ├── sprite.png
│       │       └── tensors.tsv
│       ├── 00001
│       │   └── noMetadata
│       │       └── tensors.tsv
│       ├── 00002
│       │   └── default
│       │       ├── metadata.tsv
│       │       ├── sprite.png
│       │       └── tensors.tsv
│       ├── data
│       │   └── scalar_group
│       │       ├── arctanx
│       │       │   └── events.out.tfevents.1596589834.twhuang-mbp
│       │       ├── xcosx
│       │       │   └── events.out.tfevents.1596589834.twhuang-mbp
│       │       └── xsinx
│       │           └── events.out.tfevents.1596589834.twhuang-mbp
│       ├── events.out.tfevents.1596589834.twhuang-mbp
│       └── projector_config.pbtxt
├── b
├── c
├── d
├── e
```
Where b, c, d, e have similar structure like `a`
</p>
</details>



cc @cryptopic